+++ backend/actions/sitescripts/scripts.php
@@ -2,7 +2,7 @@
 
 // All global scripts of the site
 $output = '';
-foreach ($servant->site()->scripts('server') as $file) {
+foreach ($servant->assets()->scripts('server') as $file) {
 	$output .= file_get_contents($file);
 }