+++ changelog.md
@@ -14,3 +14,12 @@ Some changes to keys used in manifest:
 - `assets` is **gone**, replaced by `scripts` and `stylesheets`
 
 All keys are treated as case-insensitive, and both plural and singular forms are accepted.
+
+
+### Upgrade guide
+
+- Update `settings.json` to use the new keys and shared setting items
+- If you've hacked `html` template, migrate changes to the new version
+- If you're using `js-vars`, some variables have changed
+
+As always, import your custom assets, pages, templates and actions.